+++ Properties/Settings.Designer.cs
@@ -297,5 +297,29 @@ namespace Wrok.Properties {
                 this["SettingsUpgraded"] = value;
             }
         }
+
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("True")]
+        public bool CheckForUpdates {
+            get {
+                return ((bool)(this["CheckForUpdates"]));
+            }
+            set {
+                this["CheckForUpdates"] = value;
+            }
+        }
+
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("")]
+        public string LastUpdateCheckUtc {
+            get {
+                return ((string)(this["LastUpdateCheckUtc"]));
+            }
+            set {
+                this["LastUpdateCheckUtc"] = value;
+            }
+        }
     }
 }